refactor: update video section
